30/12

+++ db/migrate/20161118123918_create_hocphans.rb
@@ -3,9 +3,9 @@ class CreateHocphans < ActiveRecord::Migration[5.0]
     create_table :hocphans do |t|
       t.string :mahocphan
       t.string :tenhocphan
-      t.integer :tinchi
-      t.integer :tinchihocphi
-      t.integer :trongso
+      t.float :tinchi
+      t.float :tinchihocphi
+      t.float :trongso
       t.boolean :modangki
       t.references :khoavien, foreign_key: true      
     end
